update the existing moves and add weighted centroid move generator

+++ vpr/src/place/median_move_generator.h
@@ -5,7 +5,7 @@
 
 class MedianMoveGenerator : public MoveGenerator {
     e_create_move propose_move(t_pl_blocks_to_be_moved& affected_blocks, float,
-     std::vector<float>& X_coord, std::vector<float>& Y_coord);
+     std::vector<int>& X_coord, std::vector<int>& Y_coord);
 };
 
 #endif
