+++ vpr/src/place/place.cpp
@@ -259,8 +259,8 @@ static e_move_result try_swap(const t_annealing_state* state,
                               float rlim_escape_fraction,
                               const t_place_algorithm& place_algorithm,
                               float timing_tradeoff,
-                              std::vector<float>& X_coord,
-                              std::vector<float>& Y_coord);
+                              std::vector<int>& X_coord,
+                              std::vector<int>& Y_coord);
 
 static void check_place(const t_placer_costs& costs,
                         const PlaceDelayModel* delay_model,
@@ -286,8 +286,8 @@ static float starting_t(const t_annealing_state* state,
                         ClusteredPinTimingInvalidator* pin_timing_invalidator,
                         t_pl_blocks_to_be_moved& blocks_affected,
                         const t_placer_opts& placer_opts,
-                        std::vector<float>& X_coord,
-                        std::vector<float>& Y_coord);
+                        std::vector<int>& X_coord,
+                        std::vector<int>& Y_coord);
 
 static int count_connections();
 
@@ -624,6 +624,7 @@ void try_place(const t_placer_opts& placer_opts,
     /* Set the temperature high so essentially all swaps will be accepted   */
     /* when trying to determine the starting temp for placement inner loop. */
     first_t = HUGE_POSITIVE_FLOAT;
+    
 
     t_annealing_state state(annealing_sched, first_t, first_rlim, first_move_lim, first_crit_exponent);
 
@@ -1064,8 +1065,8 @@ static float starting_t(const t_annealing_state* state,
                         ClusteredPinTimingInvalidator* pin_timing_invalidator,
                         t_pl_blocks_to_be_moved& blocks_affected,
                         const t_placer_opts& placer_opts,
-                        std::vector<float>& X_coord,
-                        std::vector<float>& Y_coord) {
+                        std::vector<int>& X_coord,
+                        std::vector<int>& Y_coord) {
     /* Finds the starting temperature (hot condition).              */
 
     int i, num_accepted, move_lim;
@@ -1191,8 +1192,8 @@ static e_move_result try_swap(const t_annealing_state* state,
                               float rlim_escape_fraction,
                               const t_place_algorithm& place_algorithm,
                               float timing_tradeoff,
-                              std::vector<float>& X_coord,
-                              std::vector<float>& Y_coord) {
+                              std::vector<int>& X_coord,
+                              std::vector<int>& Y_coord) {
     /* Picks some block and moves it to another spot.  If this spot is   *
      * occupied, switch the blocks.  Assess the change in cost function. *
      * rlim is the range limiter.                                        *
