+++ vpr/src/place/uniform_move_generator.h
@@ -4,7 +4,7 @@
 
 class UniformMoveGenerator : public MoveGenerator {
     e_create_move propose_move(t_pl_blocks_to_be_moved& affected_blocks, float rlim,
-    	std::vector<float>& , std::vector<float>& );
+    	std::vector<int>& , std::vector<int>& );
 };
 
 #endif
